Don't synchronized() inside the idle check callback
If we synchronized() against the scheduler here and we grab hold of the
lock, at the same time in another thread a callback from Mesos comes in
and that too also calls synchronized(). This behaviour casuses the Mesos
Scheduler driver to lock up because it's single threaded.

In the event that the former then decides to kill a mesos task, we'll see
a deadlock because the killTask() message can't be sent while the driver
is waiting on another callback.